Replace base64 dependency with cl-base64
Quicklisp has four systems with 'base64' in the name, the most popular
of which is cl-base64, which has 56 immediate dependencies. The base64
system has a package collision with cl-base64: the name of the former's
package and the nickname of the latter's is both base64.

This commit replaces base64 with cl-base64, so that the crud system can
be loaded on the same image as the system, its 56 immediate
dependencies, and all its dependencies thereafter.